Changed HouseHoldInfo.cars from Double to Int (#1747)

+++ src/main/scala/beam/utils/scenario/urbansim/DataExchange.scala
@@ -20,5 +20,5 @@ private[urbansim] object DataExchange {
     mode: Option[String]
   )
 
-  case class HouseholdInfo(householdId: String, cars: Double, income: Double, unitId: String, buildingId: String)
+  case class HouseholdInfo(householdId: String, cars: Int, income: Double, unitId: String, buildingId: String)
 }
